Rename and move fragment delegate classes
